+++ examples/respawn/src/main.rs
@@ -87,7 +87,8 @@ fn app_crash_dialog(args: zng::app::crash_handler::CrashArgs) {
             start_position = zng::window::StartPosition::CenterMonitor;
             auto_size = window::AutoSize::CONTENT;
             min_size = (300, 100);
-            enabled_state_cmds = !(window::WindowStateCmd::MAXIMIZE & window::WindowStateCmd::FULLSCREEN);
+            can_maximize = false;
+            can_fullscreen = false;
 
             on_load = hn_once!(|_| {
                 // force to foreground